WIP: Moved web_utils.py to this repository from volttron-testing and separated web authentication unit tests from integration tests.
